Move pyanaconda.platform
Move platform to the Storage module.
