+++ pyanaconda/modules/storage/partitioning/custom/custom_partitioning.py
@@ -35,7 +35,7 @@ from pyanaconda.modules.storage.partitioning.automatic.noninteractive_partitioni
     NonInteractivePartitioningTask
 from pyanaconda.modules.storage.partitioning.automatic.utils import get_candidate_disks, \
     schedule_partitions
-from pyanaconda.platform import platform
+from pyanaconda.modules.storage.platform import platform
 from pyanaconda.storage.utils import get_pbkdf_args, lookup_alias
 from pyanaconda.core.storage import suggest_swap_size
