+++ pyanaconda/modules/storage/platform.py
