+++ tests/nosetests/pyanaconda_tests/module_scheduler_test.py
@@ -42,7 +42,7 @@ from pyanaconda.modules.storage.partitioning.interactive.scheduler_interface imp
     DeviceTreeSchedulerInterface
 from pyanaconda.modules.storage.partitioning.interactive.scheduler_module import \
     DeviceTreeSchedulerModule
-from pyanaconda.platform import EFI
+from pyanaconda.modules.storage.platform import EFI
 from pyanaconda.storage.initialization import create_storage
 from pyanaconda.storage.root import Root
 from tests.nosetests.pyanaconda_tests import patch_dbus_publish_object, check_task_creation, \
